[imap] Gracefully handle bodystructure where boundaries are reported as NIL.
Some servers (e.g mail.ru) report boundaries as NIL for multipart messages.
from my understanding of RFC2046 this is invalid, but we can gracefully handle
those by setting them to empty.
Avoids crash when requesting a partAt() for a message that does not exist,
this happened due to NIL boundaries, but no need to crash in such cases.